Refactored AzureBotUserAuthorization to include AgentUserAuthorization internal class for simplicity
