+++ src/libraries/Builder/Microsoft.Agents.Builder/UserAuth/TokenService/AzureBotUserAuthorization.cs
@@ -17,8 +17,14 @@ namespace Microsoft.Agents.Builder.UserAuth.TokenService
     /// </summary>
     public class AzureBotUserAuthorization : OBOExchange, IUserAuthorization
     {
+        // OAuthFlow is used here as a path for back-compat since Dialog.OAuthPrompt uses
+        // it.  Long term, OAuthPrompt should probably migrate to the higher level classes
+        // and this impl refactored.
+        private readonly OAuthFlow _flow;
+
         private readonly OAuthSettings _settings;
-        private readonly AgentUserAuthorization _agentAuthentication;
+        private readonly IStorage _storage;
+        private readonly Deduplicate _dedupe;
 
         public AzureBotUserAuthorization(string name, IStorage storage, IConnections connections, IConfigurationSection configurationSection)
             : this(name, storage, connections, GetOAuthSettings(configurationSection))
@@ -38,7 +44,9 @@ namespace Microsoft.Agents.Builder.UserAuth.TokenService
         {
             Name = name ?? throw new ArgumentNullException(nameof(name));
             _settings = settings ?? throw new ArgumentNullException(nameof(settings));
-            _agentAuthentication = new AgentUserAuthorization(name, settings, storage);
+            _storage = storage ?? throw new ArgumentNullException(nameof(storage));
+            _flow = new OAuthFlow(settings);
+            _dedupe = new Deduplicate(_storage);
         }
 
         public string Name { get; private set; }
@@ -51,9 +59,9 @@ namespace Microsoft.Agents.Builder.UserAuth.TokenService
         /// <inheritdoc/>
         public async Task<TokenResponse> SignInUserAsync(ITurnContext turnContext, bool forceSignIn = false, string exchangeConnection = null, IList<string> exchangeScopes = null, CancellationToken cancellationToken = default)
         {
-            if (_agentAuthentication != null && (forceSignIn || await _agentAuthentication.IsValidActivity(turnContext, cancellationToken).ConfigureAwait(false)))
+            if (forceSignIn || await IsValidActivity(turnContext, cancellationToken).ConfigureAwait(false))
             {
-                var token = await _agentAuthentication.OnFlowTurn(turnContext, cancellationToken).ConfigureAwait(false);
+                var token = await OnFlowTurn(turnContext, cancellationToken).ConfigureAwait(false);
 
                 try
                 {
@@ -72,7 +80,7 @@ namespace Microsoft.Agents.Builder.UserAuth.TokenService
         /// <inheritdoc/>
         public async Task ResetStateAsync(ITurnContext turnContext, CancellationToken cancellationToken = default)
         {
-            await _agentAuthentication.ResetStateAsync(turnContext, cancellationToken).ConfigureAwait(false);
+            await _storage.DeleteAsync([GetStorageKey(turnContext)], cancellationToken).ConfigureAwait(false);
         }
 
         /// <inheritdoc/>
@@ -105,5 +113,155 @@ namespace Microsoft.Agents.Builder.UserAuth.TokenService
 
             return settings;
         }
+
+        private async Task<bool> IsValidActivity(ITurnContext context, CancellationToken cancellationToken = default)
+        {
+            // Catch user input in Teams where the flow has timed out.  Otherwise we get stuck in "flow active" forever.
+            if (context.Activity.ChannelId.IsParentChannel(Channels.Msteams) && context.Activity.IsType(ActivityTypes.Message))
+            {
+                var state = await GetFlowStateAsync(context, cancellationToken).ConfigureAwait(false);
+                if (state.FlowStarted && OAuthFlow.HasTimedOut(context, state.FlowExpires))
+                {
+                    throw new AuthException("Authorization flow timed out.", AuthExceptionReason.Timeout);
+                }
+            }
+
+            var isMatch = context.Activity.IsType(ActivityTypes.Message);
+
+            isMatch |= context.Activity.IsType(ActivityTypes.Invoke) &&
+                context.Activity.Name == SignInConstants.VerifyStateOperationName;
+
+            isMatch |= context.Activity.IsType(ActivityTypes.Invoke) &&
+                context.Activity.Name == SignInConstants.TokenExchangeOperationName;
+
+            isMatch |= context.Activity.IsType(ActivityTypes.Invoke) &&
+                context.Activity.Name == SignInConstants.SignInFailure;
+
+            return isMatch;
+        }
+
+        private async Task<TokenResponse> OnFlowTurn(ITurnContext turnContext, CancellationToken cancellationToken)
+        {
+            if (_settings.EnableSso && !await _dedupe.ProceedWithExchangeAsync(turnContext, cancellationToken).ConfigureAwait(false))
+            {
+                throw new DuplicateExchangeException();
+            }
+
+            var state = await GetFlowStateAsync(turnContext, cancellationToken).ConfigureAwait(false);
+
+            // Handle start or continue of the flow.
+            // Either path can throw.  This is intentionally not trapping the exception to give the caller the chance
+            // to determine if a retry is applicable.  Otherwise, caller should call ResetState.
+            TokenResponse tokenResponse;
+            if (!state.FlowStarted)
+            {
+                // If the user is already signed in, tokenResponse will be non-null
+                tokenResponse = await OnGetOrStartFlowAsync(turnContext, state, cancellationToken).ConfigureAwait(false);
+            }
+            else
+            {
+                // For non-Teams Agents, the user sends the "magic code" that will be used to exchange for a token.
+                tokenResponse = await OnContinueFlow(turnContext, state, cancellationToken);
+            }
+
+            await SaveFlowStateAsync(turnContext, state, cancellationToken).ConfigureAwait(false);
+
+            return tokenResponse;
+        }
+
+        private async Task<TokenResponse> OnGetOrStartFlowAsync(ITurnContext turnContext, FlowState state, CancellationToken cancellationToken)
+        {
+            // If the user is already signed in, tokenResponse will be non-null
+            var tokenResponse = await _flow.BeginFlowAsync(
+                turnContext,
+                null,
+                cancellationToken).ConfigureAwait(false);
+
+            // If a TokenResponse is returned, there was a cached token already.  Otherwise, start the process of getting a new token.
+            if (tokenResponse == null)
+            {
+                var expires = DateTime.UtcNow.AddMilliseconds(_settings.Timeout ?? OAuthSettings.DefaultTimeoutValue.TotalMilliseconds);
+
+                state.FlowStarted = true;
+                state.FlowExpires = expires;
+            }
+
+            return tokenResponse;
+        }
+
+        private async Task<TokenResponse> OnContinueFlow(ITurnContext turnContext, FlowState state, CancellationToken cancellationToken)
+        {
+            TokenResponse tokenResponse;
+
+            try
+            {
+                tokenResponse = await _flow.ContinueFlowAsync(turnContext, state.FlowExpires, cancellationToken).ConfigureAwait(false);
+            }
+            catch (TimeoutException)
+            {
+                throw new AuthException("Authorization flow timed out.", AuthExceptionReason.Timeout);
+            }
+            catch (UserCancelledException)
+            {
+                throw new AuthException("User cancelled authorization", AuthExceptionReason.UserCancelled);
+            }
+            catch (ConsentRequiredException)
+            {
+                await _dedupe.DeleteTokenExchangeAsync(turnContext);
+                return null;
+            }
+
+            if (tokenResponse == null)
+            {
+                if (!OAuthFlow.IsTokenExchangeRequestInvoke(turnContext))
+                {
+                    state.ContinueCount++;
+                    if (state.ContinueCount >= _settings.InvalidSignInRetryMax)
+                    {
+                        // The only way this happens is if C2 sent a bogus code
+                        throw new AuthException("Retry max", AuthExceptionReason.InvalidSignIn);
+                    }
+
+                    await turnContext.SendActivityAsync(_settings.InvalidSignInRetryMessage, cancellationToken: cancellationToken).ConfigureAwait(false);
+                }
+
+                // token still pending.
+                return null;
+            }
+
+            state.FlowStarted = false;
+            return tokenResponse;
+        }
+
+        private async Task<FlowState> GetFlowStateAsync(ITurnContext turnContext, CancellationToken cancellationToken)
+        {
+            var key = GetStorageKey(turnContext);
+            var items = await _storage.ReadAsync([key], cancellationToken).ConfigureAwait(false);
+            return items.TryGetValue(key, out object value) ? (FlowState)value : new FlowState();
+        }
+
+        private async Task SaveFlowStateAsync(ITurnContext turnContext, FlowState state, CancellationToken cancellationToken)
+        {
+            var key = GetStorageKey(turnContext);
+            var items = new Dictionary<string, object>()
+                {
+                    { key, state }
+                };
+            await _storage.WriteAsync(items, cancellationToken).ConfigureAwait(false);
+        }
+
+        private string GetStorageKey(ITurnContext turnContext)
+        {
+            var channelId = turnContext.Activity.ChannelId ?? throw new InvalidOperationException("invalid activity-missing channelId");
+            var conversationId = turnContext.Activity.Conversation?.Id ?? throw new InvalidOperationException("invalid activity-missing Conversation.Id");
+            return $"oauth/{Name}/{channelId}/{conversationId}/flowState";
+        }
+    }
+
+    class FlowState
+    {
+        public bool FlowStarted = false;
+        public DateTime FlowExpires = DateTime.MinValue;
+        public int ContinueCount = 0;
     }
 }
