+++ src/libraries/Core/Microsoft.Agents.Core/Models/TokenRequest.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using System.Collections.Generic;
 
 namespace Microsoft.Agents.Core.Models
@@ -8,6 +9,7 @@ namespace Microsoft.Agents.Core.Models
     /// <summary>
     /// A request to receive a user token.
     /// </summary>
+    [Obsolete("No longer used")]
     public class TokenRequest
     {
         public TokenRequest() { }
